feat: adds `alias` capability to usernames
this is a one-time action, which can be performed by calling `users.alias` with

{
  "username": "example@me.com",
  "alias": "zorro"
}

User must be active, not locked, have no existing alias and alias must not already be taken.
Errors with codes 412, 423, 417 and 409 would be returned if any of these conditions fail in the respective order.

When action succeeds, alias is written to hashmap of alias:username, username is added to the pool of public users,
alias is written into general `data` holder of the user and into metadata